+++ src/vue/src/flextable.js
@@ -15,7 +15,7 @@ import VueI18n from 'vue-i18n'
 
 import Flextable from './Flextable.vue'
 
-import store from './store/table'
+import store from './store/store'
 import messages from './translations/all.js'
 
 Vue.use(VueI18n)
@@ -26,12 +26,13 @@ const i18n = new VueI18n({
   messages,
 })
 
+window.initVue = (id) => {
 
-new Vue({
-  el: '.vue-flextable-app',
-  components: {
-    Flextable
-  },
-  i18n, store
-})
-
+  new Vue({
+    el: `#${id}`,
+    components: {
+      Flextable
+    },
+    i18n, store
+  })
+}