+++ CHANGELOG.md
@@ -8,3 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Converted to Craft 3
 - Rewrote vanilla JS to Vue JS
+
+## 2.0.3 - 2018-08-23
+### Changed
+- Fixed issue with flextable in matrix fields
+- Fixed issue with multiple flextables on the same page